Split shared file drop support
